Move client tool functions into new file.
